Add New device from Login
Now user will redirect to Add device from welcome screen
