+++ Yona/Yona/en.lproj/Localizable.strings
@@ -54,7 +54,7 @@
 "account-login" = "Account be created";
 "account-login-security-message" = "For security";
 "send-otp-again" = "Send code again";
-"add-device" = "Add device";
+"add-device" = "ADD DEVICE";
 "yona-add-device-message" = "Enter this passcode on the new device where you like to use Yona.";
 "loggedin-add-device-message" = "You arent logged on this device before. Go to your primary device (probably your mobile)";
 "passcode" = "CODE";
